# Inline state machine