--- a/chkjson.go
+++ b/chkjson.go
@@ -17,16 +17,6 @@
 // alternatives to encoding/json for a few specific use cases.
 package chkjson
 
-// Valid returns whether b is valid JSON.
-func Valid(b []byte) bool {
-	p := parser{in: b}
-	_, ok := p.peekAfterSpace()
-	if !ok {
-		return false
-	}
-	return p.validate()
-}
-
 type parseState byte
 
 const (
@@ -41,139 +31,36 @@ const (
 // at least deal with) is not so deep, which allows us to never allocate an
 // extension slice.
 type parseStateStack struct {
-	stackBase [32]parseState
-	baseEnd   uint8
-	stackExt  []parseState
+	base [32]parseState
+	end  uint8
+	ext  []parseState
 }
 
-func (p *parseStateStack) push(s parseState) {
-	if p.baseEnd < 32 {
-		p.stackBase[p.baseEnd] = s
-		p.baseEnd++
-		return
-	}
-	p.stackExt = append(p.stackExt, s)
-}
-
-func (p *parseStateStack) replace(s parseState) {
-	l := len(p.stackExt)
+func (p *parseStateStack) pop() parseState { // faster to not inline in endVal
+	l := len(p.ext)
 	if l == 0 {
-		p.stackBase[p.baseEnd-1] = s
-		return
-	}
-
-	p.stackExt[l-1] = s
-}
-
-func (p *parseStateStack) pop() parseState {
-	l := len(p.stackExt)
-	if l == 0 {
-		r := p.stackBase[p.baseEnd-1]
-		p.baseEnd--
+		r := p.base[p.end-1]
+		p.end--
 		return r
 	}
 
-	r := p.stackExt[l-1]
-	p.stackExt = p.stackExt[:l-1]
+	r := p.ext[l-1]
+	p.ext = p.ext[:l-1]
 	return r
 }
 
 func (p *parseStateStack) empty() bool {
-	return p.baseEnd == 0
+	return p.end == 0
 }
 
 type parser struct {
 	in []byte
 	at int
 
-	parseState parseStateStack
+	stack parseStateStack
 }
 
-func (p *parser) done() bool { return p.at == len(p.in) }
-func (p *parser) on() byte   { return p.in[p.at] }
-func (p *parser) skip()      { p.at++ }
-
-func (p *parser) peek() (byte, bool) {
-	if p.at < len(p.in) {
-		return p.in[p.at], true
-	}
-	return 0, false
-}
-
-func (p *parser) peek2() (byte, byte, bool) {
-	if p.at+1 < len(p.in) {
-		return p.in[p.at], p.in[p.at+1], true
-	}
-	return 0, 0, false
-}
-
-func (p *parser) next4() (byte, byte, byte, byte, bool) {
-	if p.at+3 < len(p.in) {
-		c1, c2, c3, c4 := p.in[p.at], p.in[p.at+1], p.in[p.at+2], p.in[p.at+3]
-		p.at += 4
-		return c1, c2, c3, c4, true
-	}
-	return 0, 0, 0, 0, false
-}
-
-func (p *parser) next() (byte, bool) {
-	if p.at < len(p.in) {
-		c := p.in[p.at]
-		p.at++
-		return c, true
-	}
-	return 0, false
-}
-
-func (p *parser) try3(c1, c2, c3 byte) bool {
-	r := p.at+2 < len(p.in) &&
-		p.in[p.at] == c1 &&
-		p.in[p.at+1] == c2 &&
-		p.in[p.at+2] == c3
-	p.at += 3
-	return r
-}
-
-func (p *parser) try4(c1, c2, c3, c4 byte) bool {
-	r := p.at+3 < len(p.in) &&
-		p.in[p.at] == c1 &&
-		p.in[p.at+1] == c2 &&
-		p.in[p.at+2] == c3 &&
-		p.in[p.at+3] == c4
-	p.at += 4
-	return r
-}
-
-func isSpace(c byte) bool {
-	return c <= ' ' && (c == ' ' || c == '\t' || c == '\r' || c == '\n')
-}
-
-func (p *parser) afterSpace() (byte, bool) {
-	for {
-		c, ok := p.next()
-		if !ok {
-			return 0, false
-		}
-		if !isSpace(c) {
-			return c, true
-		}
-	}
-}
-
-func (p *parser) peekAfterSpace() (byte, bool) {
-	for {
-		c, ok := p.peek()
-		if !ok {
-			return 0, false
-		}
-		if !isSpace(c) {
-			return c, true
-		}
-		p.skip()
-	}
-}
-
-// validate validates JSON starting from the knowledge that some non-space
+// Valid validates JSON starting from the knowledge that some non-space
 // character exists in the parser.
 //
 // This validation is modelled after the stdlib's validation but operates
@@ -191,109 +78,337 @@ func (p *parser) peekAfterSpace() (byte, bool) {
 // allows us to avoid setting the next function to use for evaluating. Every
 // "next function" is the beginning of a value. Whenever we begin a value, we
 // know what the rest of it must parse as.
-func (p *parser) validate() bool {
+//
+// We use a giant state-machine function and inline many functions for
+// significant performance gain (on the order of 50% faster).
+
+// Valid returns whether b is valid JSON.
+func Valid(b []byte) bool {
+	p := parser{in: b}
 	var c byte
-	ok := true
-	for ok {
-		if c, ok = p.afterSpace(); !ok {
-			return p.parseState.empty()
+	for p.at < len(p.in) {
+		c = p.in[p.at]
+		if !isSpace(c) {
+			goto start
 		}
-		ok = p.beginVal(c)
+		p.at++
 	}
 	return false
-}
 
-func (p *parser) beginVal(c byte) bool {
+start:
+	for p.at < len(p.in) { // afterSpace
+		c = p.in[p.at]
+		p.at++
+		if !isSpace(c) {
+			goto beginVal
+		}
+	}
+	return p.stack.empty()
+
+beginVal:
 	switch c {
 	case '{':
-		p.parseState.push(parseObjKey)
-		return p.beginStrOrEmpty()
+		if p.stack.end < 32 { // push
+			p.stack.base[p.stack.end] = parseObjKey
+			p.stack.end++
+		} else {
+			p.stack.ext = append(p.stack.ext, parseObjKey)
+		}
+		goto beginStrOrEmpty
 
 	case '[':
-		p.parseState.push(parseArrVal)
-		return p.beginValOrEmpty()
+		if p.stack.end < 32 { // push
+			p.stack.base[p.stack.end] = parseArrVal
+			p.stack.end++
+		} else {
+			p.stack.ext = append(p.stack.ext, parseArrVal)
+		}
+		goto beginValOrEmpty
 
 	case '"':
-		return p.beganStrFin()
+		goto finStr
 
 	case '-':
-		return p.finNeg()
+		goto finNeg
 
 	case '0':
-		return p.fin0()
+		goto fin0
 
 	case 't':
-		return p.finTrue()
+		goto finTrue
 
 	case 'f':
-		return p.finFalse()
+		goto finFalse
 
 	case 'n':
-		return p.finNull()
+		goto finNull
 	}
 	if isNat(c) {
-		return p.fin1()
+		goto fin1
 	}
 	return false
-}
 
-func (p *parser) beginValOrEmpty() bool {
-	c, ok := p.peekAfterSpace()
-	return ok && (c != ']' || p.endVal())
-}
-
-func (p *parser) beginStrOrEmpty() bool {
-	c, ok := p.peekAfterSpace()
-	if !ok {
-		return false
+beginStrOrEmpty:
+	for p.at < len(p.in) { // afterSpace
+		c = p.in[p.at]
+		p.at++
+		if !isSpace(c) {
+			break
+		}
 	}
 	if c == '}' {
-		p.parseState.replace(parseObjVal)
-		return p.endVal()
-	}
-	p.skip()
-	return c == '"' && p.finStr() && p.endVal()
-}
-
-// beganStrFin finishes a started string and falls into endVal.
-func (p *parser) beganStrFin() bool {
-	return p.finStr() && p.endVal()
-}
-
-func (p *parser) finStr() bool {
-start:
-	c, ok := p.next()
-	if !ok {
-		return false
+		l := len(p.stack.ext)
+		if l == 0 {
+			p.stack.end--
+		} else {
+			p.stack.ext = p.stack.ext[:l-1]
+		}
+		goto endVal
 	}
 	if c == '"' {
-		return true
+		goto finStr
 	}
-	if c == '\\' {
-		if ok = p.finStrEsc(); !ok {
+	return false
+
+beginValOrEmpty:
+	for p.at < len(p.in) {
+		c = p.in[p.at]
+		if !isSpace(c) {
+			if c == ']' {
+				goto endVal
+			}
+			goto start
+		}
+		p.at++
+	}
+	return false
+
+finStr:
+	if p.at >= len(p.in) {
+		return false
+	}
+	c = p.in[p.at]
+	p.at++
+	if c == '"' {
+		goto endVal
+	}
+	if c == '\\' { // finStrEsc
+		if p.at >= len(p.in) {
 			return false
 		}
-		goto start
+		c = p.in[p.at]
+		p.at++
+		switch c {
+		case 'b', 'f', 'n', 'r', 't', '\\', '/', '"':
+			goto finStr
+		case 'u':
+			if p.at+3 < len(p.in) &&
+				isHex(p.in[p.at]) &&
+				isHex(p.in[p.at+1]) &&
+				isHex(p.in[p.at+2]) &&
+				isHex(p.in[p.at+3]) {
+				p.at += 4
+				goto finStr
+			}
+		}
+		return false
 	}
 	if c < 0x20 {
 		return false
 	}
-	goto start
-}
+	goto finStr
 
-func (p *parser) finStrEsc() bool {
-	c, ok := p.next()
-	if !ok {
+finNeg:
+	if p.at >= len(p.in) {
 		return false
 	}
-	switch c {
-	case 'b', 'f', 'n', 'r', 't', '\\', '/', '"':
-		return true
-	case 'u':
-		c1, c2, c3, c4, ok := p.next4()
-		return ok && isHex(c1) && isHex(c2) && isHex(c3) && isHex(c4)
+	c = p.in[p.at]
+	p.at++
+	if c == '0' {
+		goto fin0
+	}
+	if isNat(c) {
+		goto fin1
 	}
 	return false
+
+fin1:
+	for p.at < len(p.in) {
+		if !isNum(p.in[p.at]) {
+			goto fin0
+		}
+		p.at++
+	}
+	goto endVal
+
+fin0:
+	if p.at >= len(p.in) {
+		goto endVal
+	}
+	c = p.in[p.at]
+	if c == '.' {
+		p.at++
+		goto finDot
+	}
+	if isE(c) {
+		p.at++
+		goto finE
+	}
+	goto endVal
+
+finDot:
+	if p.at >= len(p.in) {
+		return false
+	}
+	c = p.in[p.at]
+	p.at++
+	if !isNum(c) { // first char after dot must be num
+		return false
+	}
+
+	for p.at < len(p.in) && isNum(p.in[p.at]) { // consume all nums
+		p.at++
+	}
+
+	if p.at < len(p.in) && isE(p.in[p.at]) {
+		p.at++
+		goto finE
+	}
+	goto endVal
+
+finE:
+	if p.at >= len(p.in) {
+		return false
+	}
+	c = p.in[p.at]
+	p.at++
+	if c == '+' || c == '-' {
+		if p.at >= len(p.in) {
+			return false
+		}
+		c = p.in[p.at]
+		p.at++
+	}
+	if !isNum(c) { // first after e (and +/-) must be num
+		return false
+	}
+	for p.at < len(p.in) && isNum(p.in[p.at]) { // consume all nums
+		p.at++
+	}
+	goto endVal
+
+finTrue:
+	if p.at+2 < len(p.in) &&
+		p.in[p.at] == 'r' &&
+		p.in[p.at+1] == 'u' &&
+		p.in[p.at+2] == 'e' {
+
+		p.at += 3
+		goto endVal
+	}
+	return false
+
+finFalse:
+	if p.at+3 < len(p.in) &&
+		p.in[p.at] == 'a' &&
+		p.in[p.at+1] == 'l' &&
+		p.in[p.at+2] == 's' &&
+		p.in[p.at+3] == 'e' {
+
+		p.at += 4
+		goto endVal
+	}
+	return false
+
+finNull:
+	if p.at+2 < len(p.in) &&
+		p.in[p.at] == 'u' &&
+		p.in[p.at+1] == 'l' &&
+		p.in[p.at+2] == 'l' {
+
+		p.at += 3
+		goto endVal
+	}
+	return false
+
+endVal: // most things parsed fall into endVal
+	if p.stack.empty() {
+		for p.at < len(p.in) { // afterSpace
+			c = p.in[p.at]
+			p.at++
+			if !isSpace(c) {
+				return false
+			}
+		}
+		return true // the rest are space if we have nothing after spaces
+	}
+
+	for p.at < len(p.in) { // if parseState is not empty, we need another character
+		c = p.in[p.at]
+		p.at++
+		if !isSpace(c) {
+			goto finVal
+		}
+	}
+	return false
+
+finVal:
+	switch p.stack.pop() {
+	case parseObjKey:
+		if c == ':' {
+			if p.stack.end < 32 { // push
+				p.stack.base[p.stack.end] = parseObjVal
+				p.stack.end++
+			} else {
+				p.stack.ext = append(p.stack.ext, parseObjVal)
+			}
+			goto start
+		}
+
+	case parseObjVal:
+		switch c {
+		case ',':
+			if p.stack.end < 32 { // push
+				p.stack.base[p.stack.end] = parseObjKey
+				p.stack.end++
+			} else {
+				p.stack.ext = append(p.stack.ext, parseObjKey)
+			}
+			for p.at < len(p.in) { // afterSpace
+				c = p.in[p.at]
+				p.at++
+				if isSpace(c) {
+					continue
+				}
+				if c == '"' {
+					goto finStr
+				}
+				return false
+			}
+		case '}':
+			goto endVal
+		}
+
+	case parseArrVal:
+		switch c {
+		case ',':
+			if p.stack.end < 32 { // push
+				p.stack.base[p.stack.end] = parseArrVal
+				p.stack.end++
+			} else {
+				p.stack.ext = append(p.stack.ext, parseArrVal)
+			}
+			goto start
+		case ']':
+			goto endVal
+		}
+	}
+
+	return false
+}
+
+func isSpace(c byte) bool {
+	return c <= ' ' && (c == ' ' || c == '\t' || c == '\r' || c == '\n')
 }
 
 func isHex(c byte) bool {
@@ -310,140 +425,4 @@ func isNat(c byte) bool {
 
 func isE(c byte) bool {
 	return c == 'e' || c == 'E'
-}
-
-func (p *parser) finNeg() bool {
-	c, ok := p.next()
-	return ok && (c == '0' && p.fin0() || isNat(c) && p.fin1())
-}
-
-func (p *parser) fin1() bool {
-	for !p.done() {
-		if !isNum(p.on()) {
-			return p.fin0()
-		}
-		p.skip()
-	}
-	return p.endVal()
-}
-
-func (p *parser) fin0() bool {
-	c, ok := p.peek()
-	if !ok {
-		return p.endVal()
-	}
-	if c == '.' {
-		p.skip()
-		return p.finDot()
-	}
-	if isE(c) {
-		p.skip()
-		return p.finE()
-	}
-	return p.endVal()
-}
-
-func (p *parser) finDot() bool {
-	c, ok := p.next()
-	if !ok || !isNum(c) { // first char after dot must be num
-		return false
-	}
-
-	for !p.done() && isNum(p.on()) { // consume all nums
-		p.skip()
-	}
-
-	if !p.done() && isE(p.on()) {
-		p.skip()
-		return p.finE()
-	}
-	return p.endVal()
-}
-
-func (p *parser) finE() bool {
-	c, ok := p.next()
-	if !ok {
-		return false
-	}
-	if c == '+' || c == '-' {
-		if c, ok = p.next(); !ok {
-			return false
-		}
-	}
-	if !isNum(c) { // first after e (and +/-) must be num
-		return false
-	}
-	for !p.done() && isNum(p.on()) { // consume all nums
-		p.skip()
-	}
-	return p.endVal()
-}
-
-func (p *parser) finTrue() bool {
-	return p.try3('r', 'u', 'e') && p.endVal()
-}
-
-func (p *parser) finNull() bool {
-	return p.try3('u', 'l', 'l') && p.endVal()
-}
-
-func (p *parser) finFalse() bool {
-	return p.try4('a', 'l', 's', 'e') && p.endVal()
-}
-
-// endVal is the bottom of every value. All functions called from endVal cannot
-// fall back into endVal.
-func (p *parser) endVal() bool {
-start:
-	if p.parseState.empty() {
-		return p.remSpace()
-	}
-
-	c, ok := p.afterSpace()
-	if !ok { // if parseState is not empty, we need another character
-		return false
-	}
-
-	switch p.parseState.pop() {
-	case parseObjKey:
-		if c == ':' {
-			p.parseState.push(parseObjVal)
-			return true
-		}
-
-	case parseObjVal:
-		switch c {
-		case ',':
-			p.parseState.push(parseObjKey)
-			if !p.startAndFinStr() {
-				return false
-			}
-			goto start
-		case '}':
-			goto start
-		}
-
-	case parseArrVal:
-		switch c {
-		case ',':
-			p.parseState.push(parseArrVal)
-			return true
-		case ']':
-			goto start
-		}
-	}
-
-	return false
-}
-
-func (p *parser) remSpace() bool {
-	_, ok := p.afterSpace()
-	return !ok // the rest are space if we have nothing after spaces
-}
-
-// startAndFinStr begins and finishes a string; this does not fall into endVal,
-// thus it can be called from endVal.
-func (p *parser) startAndFinStr() bool {
-	c, ok := p.afterSpace()
-	return ok && c == '"' && p.finStr()
 }