… README.md
… chkjson_test.go
… ext_test.go
… testdata/canada.json
… testdata/citm.json
… testdata/large.json
… testdata/medium.json
… testdata/small.json
… testdata/twitter.json