This switches everything to an inline state machine (one for both Valid and Compact).

The code is ugly but can be easily translated to what came before. The performance improvements are worth it.

Against master for the prior benchmarks:
```
name       old time/op    new time/op    delta
Valid-4       325ns ± 0%     182ns ± 0%  -44.00%  (p=0.000 n=9+9)
Compact-4     398ns ± 0%     219ns ± 0%  -44.88%  (p=0.000 n=8+8)
```

The README now includes benchmarks (from valyala/fastjson) for larger json blobs against the stdlib functions.